requirements.txt
fastapi
uvicorn[standard]
pyiceberg[s3fs,pyarrow,sql-sqlite]
duckdb==1.2.2
python-dotenv 